orange-pi-4.4-rockchip64-cros

+++ arch/xtensa/include/asm/checksum.h
@@ -123,8 +123,9 @@ static __inline__ __sum16 ip_fast_csum(const void *iph, unsigned int ihl)
 }
 
 static __inline__ __wsum csum_tcpudp_nofold(__be32 saddr, __be32 daddr,
-					    __u32 len, __u8 proto,
-					    __wsum sum)
+						   unsigned short len,
+						   unsigned short proto,
+						   __wsum sum)
 {
 
 #ifdef __XTENSA_EL__
@@ -156,8 +157,9 @@ static __inline__ __wsum csum_tcpudp_nofold(__be32 saddr, __be32 daddr,
  * returns a 16-bit checksum, already complemented
  */
 static __inline__ __sum16 csum_tcpudp_magic(__be32 saddr, __be32 daddr,
-					    __u32 len, __u8 proto,
-					    __wsum sum)
+						       unsigned short len,
+						       unsigned short proto,
+						       __wsum sum)
 {
 	return csum_fold(csum_tcpudp_nofold(saddr,daddr,len,proto,sum));
 }
@@ -175,7 +177,7 @@ static __inline__ __sum16 ip_compute_csum(const void *buff, int len)
 #define _HAVE_ARCH_IPV6_CSUM
 static __inline__ __sum16 csum_ipv6_magic(const struct in6_addr *saddr,
 					  const struct in6_addr *daddr,
-					  __u32 len, __u8 proto,
+					  __u32 len, unsigned short proto,
 					  __wsum sum)
 {
 	unsigned int __dummy;
